style: product 컴포넌트 styled, type 파일 분리
