+++ src/components/product/ProductField/index.tsx
@@ -1,14 +1,7 @@
-import styled from '@emotion/styled'
 import { Divider, Text } from '@offer-ui/react'
 import type { ReactElement } from 'react'
-import type { StyledProps } from '@types'
-
-export type ProductFieldProps = {
-  label: string
-  value: string
-  height?: number
-}
-type StyledTextWrapperProps = StyledProps<ProductFieldProps, 'height'>
+import { StyledProductField, StyledTextWrapper } from './styled'
+import type { ProductFieldProps } from './types'
 
 export const ProductField = ({
   label,
@@ -27,15 +20,3 @@ export const ProductField = ({
     </StyledProductField>
   )
 }
-const StyledProductField = styled.div`
-  display: flex;
-  gap: 12px;
-`
-const StyledTextWrapper = styled.div<StyledTextWrapperProps>`
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  justify-content: center;
-
-  height: ${({ height }): string => `${height}px`};
-`
