+++ src/components/product/UserProfile/index.tsx
@@ -1,23 +1,8 @@
-import styled from '@emotion/styled'
 import { Avatar, Badge, Text } from '@offer-ui/react'
 import type { ReactElement } from 'react'
-
-export const TRANSACTION_TYPE = {
-  all: '직거래/택배거래',
-  direct: '직거래',
-  parcel: '택배거래'
-} as const
-
-type TransactionType = keyof typeof TRANSACTION_TYPE
-export type UserProfileProps = {
-  image?: string
-  nickName: string
-  location: string
-  type: 'offer' | 'basic'
-  level: number
-  date?: string
-  transactionType?: TransactionType
-}
+import { StyledUserProfile, StyledProfileText, StyledUserName } from './styled'
+import type { UserProfileProps } from './types'
+import { TRANSACTION_TYPE } from './types'
 
 export const UserProfile = ({
   nickName,
@@ -56,19 +41,3 @@ export const UserProfile = ({
     </StyledUserProfile>
   )
 }
-
-const StyledUserProfile = styled.div`
-  display: flex;
-  gap: 12px;
-  align-items: center;
-`
-const StyledProfileText = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-`
-const StyledUserName = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-`
