+++ src/components/product/UserProfile/styled.ts
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled'
+
+export const StyledUserProfile = styled.div`
+  display: flex;
+  gap: 12px;
+  align-items: center;
+`
+export const StyledProfileText = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+`
+export const StyledUserName = styled.div`
+  display: flex;
+  flex-direction: row;
+  gap: 8px;
+`
